# docs/swagger.yaml
basePath: /
definitions:
...
  history-api_internal_dtos_request.CreateCommitDto:
    properties:
      edit_summary:
        maxLength: 500
        type: string
      snapshot_json:
        items:
          type: integer
        type: array
    required:
    - edit_summary
    - snapshot_json
    type: object
...
  history-api_internal_dtos_request.RestoreCommitDto:
    properties:
      commit_id:
        type: string
    required:
    - commit_id
    type: object
...
  history-api_internal_dtos_response.CommonResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for History API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: History API
  version: "1.0"
paths:
...
  /entities:
    get:
      consumes:
      - application/json
      description: Search entities with cursor pagination
      parameters:
      - in: query
        name: cursor
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maxLength: 255
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      summary: Search entities
      tags:
      - Entities
  /entities/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific entity
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      summary: Get entity by ID
      tags:
      - Entities
  /geometries:
    get:
      consumes:
      - application/json
      description: Search geometries with cursor pagination and spatial filtering
      parameters:
      - in: query
        name: entity_id
        type: string
      - in: query
        maximum: 90
        minimum: -90
        name: max_lat
        required: true
        type: number
      - in: query
        maximum: 180
        minimum: -180
        name: max_lng
        required: true
        type: number
      - in: query
        maximum: 90
        minimum: -90
        name: min_lat
        required: true
        type: number
      - in: query
        maximum: 180
        minimum: -180
        name: min_lng
        required: true
        type: number
      - in: query
        name: time
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      summary: Search geometries
      tags:
      - Geometries
  /geometries/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific geometry
      parameters:
      - description: Geometry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      summary: Get geometry by ID
      tags:
      - Geometries
...
  /projects/{id}/commits:
    get:
      consumes:
      - application/json
      description: Retrieve all commits for a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get project commits
      tags:
      - Commits
    post:
      consumes:
      - application/json
      description: Create a new commit and update project's latest commit ID. Only
        owner/editor allowed.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Commit Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.CreateCommitDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Create a new commit
      tags:
      - Commits
  /projects/{id}/commits/restore:
    post:
      consumes:
      - application/json
      description: Update project's latest commit ID to an older commit. Only owner/editor
        allowed.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Restore Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.RestoreCommitDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Restore project to a commit
      tags:
      - Commits
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"